.github/scripts/validate-parameters/requirements.txt
PyYAML ~= 6.0
